move routing_model
